California Reservoirs_MQTT/requirements.txt
paho-mqtt==1.6.1
pandas==2.2.3
matplotlib==3.8.3
apscheduler==3.10.4